Fix incorrect refactor
