+++ decidim-system/lib/decidim/system/test/factories.rb
@@ -1,3 +1,11 @@
 # frozen_string_literal: true
 
-require "decidim/system/test/factories"
+require "decidim/core/test/factories"
+
+FactoryBot.define do
+  factory :admin, class: "Decidim::System::Admin" do
+    sequence(:email) { |n| "admin#{n}@citizen.corp" }
+    password { "password1234" }
+    password_confirmation { "password1234" }
+  end
+end
