+++ decidim-system/spec/factories.rb
@@ -1,11 +1,3 @@
 # frozen_string_literal: true
 
-require "decidim/core/test/factories"
-
-FactoryBot.define do
-  factory :admin, class: "Decidim::System::Admin" do
-    sequence(:email) { |n| "admin#{n}@citizen.corp" }
-    password { "password1234" }
-    password_confirmation { "password1234" }
-  end
-end
+require "decidim/system/test/factories"
